% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Candidate-class.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{initialize,createCandidate-method}
\title{A presidential candidate object}
\description{
Object of class \code{Candidate} as created by the \code{createCandidate} function. Objects
of this class include \code{show} and \code{print} methods.
}
\details{
An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} A character string that is the name of the candidate
\item \code{delegatesWon} The number of delegates already won by the candidate
\item \code{party} The political party of the candidate
\item \code{delegatesNeeded} The number of delegates a candidate needs to win
}
}
\author{
Danielle Korman
}

