% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Create an instance of class Candidate}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{name}{A character object}

\item{delegatesWon}{A numeric object}

\item{party}{A character object}
}
\value{
An object of class Candidate containing
 \item{name}{The name of the candidate}
 \item{delegatesWon}{How many delegates the candidate has won}
 \item{party}{The political party of the candidate}
 \item{delegatesNeeded}{The number of delegates the candidate needs to win nomination}
}
\description{
Takes candidate information and creates an instance of class Candidate with delegates needed
}
\note{
This produces an instance of class Candidate
}
\examples{
B_name <- "Bernie Sanders"
B_won <- 574
B_party <- "Democrat"
createCandidate(B_name, B_won, B_party)
}
\author{
Danielle Korman
}
\seealso{
\code{\link{propNeeded}}
}

