% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propNeeded.R
\name{propNeeded}
\alias{propNeeded}
\alias{propNeeded,ANY-method}
\title{Calculates the proportion of delegates remaining needed for a Candidtate to win}
\usage{
propNeeded(candidate, delegatesRemaining)
}
\arguments{
\item{candidate}{An object of class Candidate}

\item{delegatesRemaining}{A numeric object of the number of remaining delegates for party}
}
\value{
A numeric object of the proportion of delegates remaining needed for candidate to win
}
\description{
Takes a candidate and number of delegates remaining for their party to find how many delegates
needed to win
}
\note{
Also shows if candidate has already lost
}
\examples{
Trump <- createCandidate("Donald Trump", 678, "Republican")
propNeeded(Trump, 1049)
}
\author{
Danielle Korman
}
\seealso{
\code{\link{createCandidate}}
}

